// src/interfaces/characterDetail.ts
import Episode from './episode';

interface CharacterDetail {
  id: number;
  name: string;
  status: string;
  species: string;
  type: string;
  gender: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: Episode[];
}

export default CharacterDetail;